move isauth check to session manager and enable token refresh
