+++ frontend/js/clients/token-client.js
@@ -37,7 +37,7 @@ export const register = async ({ username, email, password }) => {
 };
 
 // checks if user is authenticated (if access token is valid)
-export const isAuth = async () => {
+export const tokenStatus = async () => {
 	const url = URIS.auth;
 	const { status, body, error } = await get(url);
 	return status == 200 ? true : false;
